==> scripts/SCRIPT2_EXTRACT_PDF.ipynb <==
import PyPDF2


def extract_text_from_pdf(pdf_path):
    text = ""
    with open(pdf_path, "rb") as file:
        reader = PyPDF2.PdfReader(file)

        for page_number in range(len(reader.pages)):
            page = reader.pages[page_number]
            text += page.extract_text()

    return text

document_name='Patents_Licensing_and_IP_Policy'

extracted_text=extract_text_from_pdf(document_name+'.pdf')

with open(document_name+'.txt','w') as file:
  file.write(extracted_text)
