==> scripts/SCRIPT5_MULTIPLE_EXAMPLES.ipynb <==
import openai
import os
import pandas as pd
import time
pd.set_option('display.max_colwidth', None)

# Reading question and answer from the master spreadsheet

df = pd.read_csv("masters_spreadsheet_with_unique_intent_version3.csv")
df.head()

len(df)

def get_completion(prompt, model="gpt-3.5-turbo", retries=3):
    for _ in range(retries):
        try:
            messages = [{"role": "user", "content": prompt}]
            response = openai.ChatCompletion.create(
                model=model,
                messages=messages,
                temperature=0, # this is the degree of randomness of the model's output
            )
            return response.choices[0].message["content"]
        except openai.error.ServiceUnavailableError as e:
            print(f"Service Unavailable. Retrying after 5 seconds... Error: {e}")
            time.sleep(5)
        except openai.error.APIError as e:
            print(f"API Error. Retrying after 5 seconds... Error: {e}")
            time.sleep(5)
    raise Exception("API call failed after multiple retries.")

# Function for generating intents with examples for each question

def intents_examples(question):
    
    text = question
    
    prompt = f"""
Generate atleast ten similar short questions for the question given in the text

\"\"\"{text}\"\"\"
"""

    response = get_completion(prompt)
    return response

# Add a new column 'examples' to store the generated examples
df['examples'] = ""

# Iterate over each row in the DataFrame and generate examples for each question
for idx, row in df.iterrows():
    question = row['question']
    examples = intents_examples(question)
    df.at[idx, 'examples'] = examples

df

df.to_csv("master_spreadsheet_with_multiple_examples_version3.csv", index=False)
