==> scripts/SCRIPT6_NLU_YML_FILE.ipynb <==
# Importing libraries

import pandas as pd
pd.set_option('display.max_colwidth', None)

# Reading the master spreadsheet

df = pd.read_csv("master_spreadsheet_with_multiple_examples_version3.csv")
df.head()

len(df)

# Function to generate nlu.yml file for each row

def generate_nlu_section(df):
    nlu_content = ""
    for _, row in df.iterrows():
        question = row['question']
        unique_intent = row['unique_intents']
        examples = row['examples'].split('\n')
        
        nlu_content += f"- intent: {unique_intent}\n"
        nlu_content += "  examples: |\n"
        
        nlu_content += f"    - {question}\n"
        
        for example in examples:
            if example.strip():
                # Remove numbering (e.g., "1.", "2.", etc.) from the examples
                example_text = " ".join(example.split()[1:])
                nlu_content += f"    - {example_text.strip()}\n"
        
        nlu_content += "\n"
    
    return nlu_content

# Generate nlu.yml content
nlu_content = generate_nlu_section(df)

# Save the nlu content to nlu.yml file
with open("nlu.yml", "w") as file:
    file.write(nlu_content)
