==> scripts/SCRIPT7_STORIES_DOMAIN_YML_FILES.ipynb <==
# Importing libraries

import pandas as pd

# Reading the master spreadsheet

df = pd.read_csv("masters_spreadsheet_with_unique_intent_version3.csv")
df.head()

len(df)

# Function to generate stories.yml file for each row

def generate_stories_yaml_row(unique_intents):
    yml_content = (
        f"- story: {unique_intents}\n"
        f"  steps:\n"
        f"  - intent: {unique_intents}\n"
        f"  - action: utter_{unique_intents}\n"
    )
    return yml_content

# Convert unique intents to YAML format
yml_content = "\n".join(df.apply(lambda row: generate_stories_yaml_row(row["unique_intents"]), axis=1))

# Save the YAML content to a file
with open("stories.yml", "w") as file:
    file.write(yml_content)

# Functions to generate domain.yml file for each row

# Function to generate intents section in YAML format
def generate_intents_section(df):
    #intents = "\n  - ".join(df['unique_intents'])
    intents = "\n- ".join(df['unique_intents'])
    return f"intents:\n- {intents}\n\n"

# Function to generate actions section in YAML format
def generate_actions_section(df):
    actions = "\n- ".join([f"utter_{intent}" for intent in df['unique_intents']])
    return f"actions:\n- {actions}\n\n"

# Function to generate responses section in YAML format
def generate_responses_section(df):
    responses = "responses:\n"
    for _, row in df.iterrows():
        responses += f"  utter_{row['unique_intents']}:\n  - text: {row['answer']}\n\n"
    return responses

# Generate YAML content
yml_content = generate_intents_section(df)
yml_content += generate_actions_section(df)
yml_content += generate_responses_section(df)

# Save the YAML content to a file
with open("domain.yml", "w") as file:
    file.write(yml_content)
